requirements.txt
torch>=1.8.0
torchvision>=0.9.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
numpy>=1.20.3
pandas>=1.3.0
opencv-python>=4.5.2
streamlit>=1.0.0
pillow>=8.2.0